+++ test/guarded_struct_test/basic_types_test.exs
@@ -1,0 +1,5 @@
+defmodule MishkaDeveloperToolsTest.GuardedStruct.BasicTypesTest do
+  use ExUnit.Case, async: true
+
+  ############# (▰˘◡˘▰) BasicTypesTest GuardedStructTest Data (▰˘◡˘▰) ##############
+end