+++ test/guarded_struct_test/core_keys_test.exs
@@ -1,0 +1,5 @@
+defmodule MishkaDeveloperToolsTest.GuardedStruct.CoreKeysTest do
+  use ExUnit.Case, async: true
+
+  ############# (▰˘◡˘▰) CoreKeysTest GuardedStructTest Data (▰˘◡˘▰) ##############
+end