+++ test/guarded_struct_test/global_test.exs
@@ -1,0 +1,5 @@
+defmodule MishkaDeveloperToolsTest.GuardedStruct.GlobalTest do
+  use ExUnit.Case, async: true
+
+  ############# (▰˘◡˘▰) GlobalTest GuardedStructTest Data (▰˘◡˘▰) ##############
+end