+++ test/guarded_struct_test/sub_field_external_field_test.exs
@@ -1,0 +1,5 @@
+defmodule MishkaDeveloperToolsTest.GuardedStruct.SubFieldExternalFieldTest do
+  use ExUnit.Case, async: true
+
+  ############# (▰˘◡˘▰) SubFieldExternalFieldTest GuardedStructTest Data (▰˘◡˘▰) ##############
+end