+++ test/guarded_struct_test/validator_derive_test.exs
@@ -1,0 +1,5 @@
+defmodule MishkaDeveloperToolsTest.GuardedStruct.ValidatorDeriveTest do
+  use ExUnit.Case, async: true
+
+  ############# (▰˘◡˘▰) ValidatorDeriveTest GuardedStructTest Data (▰˘◡˘▰) ##############
+end